Fin Tache

+++ RadForm1.cs
@@ -28,34 +28,31 @@ namespace GenerateArkaveImport
             string Login = tbLogin.Text;
             string password = tbPassword.Text;
 
-            conString = $"Server = {server}; Database = CADASTRE_CONTROLE_TAZA; User ID = {Login}; Password = {password}";
+            conString = $"Server = {server}; Database = master ; User ID = {Login}; Password = {password}";
             using (SqlConnection con = new SqlConnection(conString))
             {
                 try
                 {
                     con.Open();
-                    MessageBox.Show("Connection Avec Succe");
-                    string req = "select DISTINCT id_tranche from TB_UNITE";
-                    SqlDataAdapter da = new SqlDataAdapter(req, con);
-                    DataTable dt = new DataTable();
-                    da.Fill(dt);
-                    if (dt.Rows.Count == 0)
+                    string REQ = "SELECT name FROM sys.databases where name like'%_ANCFCC'";
+                    SqlDataAdapter DA = new SqlDataAdapter(REQ, con);
+                    DataTable DT = new DataTable();
+                    DA.Fill(DT);
+                    if (DT.Rows.Count == 0)
                     {
-                        MessageBox.Show("aucun tranche");
+                        MessageBox.Show("aucune BaseDonnée");
 
                         return;
                     }
                     else
                     {
-                        comboxTranche.DataSource = dt.DefaultView;
-                        comboxTranche.DisplayMember = "id_tranche";
-                        comboxTranche.ValueMember = "id_tranche";
+                        ComboBaseDonnee.DataSource = DT;
+                        ComboBaseDonnee.DisplayMember = "name";
+                        ComboBaseDonnee.ValueMember = "name";
+                        ComboBaseDonnee.SelectedIndexChanged += ComboBaseDonnee_SelectedIndexChanged;
                     }
-                    if (comboxTranche.Items.Count > 0)
-                    {
-                        bntGenerer.Enabled = true;
-                    }
-
+                    MessageBox.Show("Connexion OK!!!");
+                    
                 }
                 catch (Exception ex)
                 {
@@ -82,8 +79,8 @@ namespace GenerateArkaveImport
                     try
                     {
                         con.Open();
-                        string id_tranche = comboxTranche.SelectedValue.ToString();
-                        string reqD = $"select top 10 * from dossier where id_unite in(select id_unite from tb_unite where id_tranche = {id_tranche})";
+                        string id_tranche = comboxTranche.Text;
+                        string reqD = $"select top 10 * from TB_Dossier where id_unite in(select id_unite from tb_unite where id_tranche = {id_tranche})";
                         SqlDataAdapter da = new SqlDataAdapter(reqD, con);
                         DataTable DT = new DataTable();
                         da.Fill(DT);
@@ -109,9 +106,7 @@ namespace GenerateArkaveImport
                     {
                         MessageBox.Show(ex.Message);
                     }
-
                 }
-
             }
         }
 
@@ -122,35 +117,42 @@ namespace GenerateArkaveImport
             foreach (DataRow doss in dT.Rows)
             {
                 string nomDoss = "";
-                string NatureOrigine = doss["nature_origine"].ToString();
+                string NatureOrigine = doss["nature_origine"].ToString().Trim();
                 string id_dossier = doss["id_dossier"].ToString();
                 if (NatureOrigine == "R")
                 {
-                    string NumeroOrigine = doss["Numero_origine"].ToString();
-                    string indiceOrigine = doss["indice_origine"].ToString();
+                    string NumeroOrigine = doss["numero_origine"].ToString().Trim();
+                    string indiceOrigine = doss["indice_origine"].ToString().Trim();
                     nomDoss = $"R{NumeroOrigine}-{indiceOrigine}";
                 }
                 else if (NatureOrigine == "T")
                 {
-                    string NumeroOrigine = doss["Numero_Titre"].ToString();
-                    string indiceOrigine = doss["indice_Titre"].ToString();
-                    nomDoss = $"R{NumeroOrigine}-{indiceOrigine}";
+                    string NumeroOrigine = doss["numero_titre"].ToString().Trim();
+                    string indiceOrigine = doss["indice_Titre"].ToString().Trim();
+                    nomDoss = $"T{NumeroOrigine}-{indiceOrigine}";
                 }
                 string dossPath = Path.Combine(path, nomDoss);
                 if (!Directory.Exists(dossPath))
                 {
                     Directory.CreateDirectory(dossPath);
                 }
-                DB dB = new DB(conString);
-                // Pieces sans hors sous dossiers
-                DataTable piecesSD0 = dB.PieceSD0(id_dossier);
-                foreach (DataRow piece in piecesSD0.Rows)
-                {
-                    string num_pg = piece["Numero_page"].ToString();
-                    string nomPiece = piece["Nature_Acte"].ToString();
-                    string chemin = piece["CHEMIN_PHYSIQUE"].ToString();
 
-                    string sdPath = Path.Combine(dossPath, "sd0");
+                DB dB = new DB(conString);
+
+                // Pieces
+                DataTable pieces = dB.Piece(id_dossier);
+                foreach (DataRow piece in pieces.Rows)
+                {
+                    string num_page = piece["Num_page"].ToString().Trim();
+                    string nomPiece = piece["nom_piece"].ToString().Trim();
+                    string chemin = piece["url"].ToString().Trim();
+
+                    int numSousDossier = Convert.ToInt32( piece["num_sous_dos"].ToString().Trim());
+                    String Formalite = piece["formalite"].ToString().Trim();
+
+                    String SousDossierName = numSousDossier != 0 ? $"sd{numSousDossier} {Formalite}" : "sd0";
+
+                    string sdPath = Path.Combine(dossPath, SousDossierName);
                     if (!Directory.Exists(sdPath))
                     {
                         Directory.CreateDirectory(sdPath);
@@ -162,51 +164,49 @@ namespace GenerateArkaveImport
                         Directory.CreateDirectory(piecePath);
                     }
 
-                    string imgNewPath = Path.Combine(piecePath, $"p{num_pg}{new FileInfo(chemin).Extension}");
-                    if (!File.Exists(imgNewPath))
+                    string imgNewPath = Path.Combine(piecePath, $"p{num_page}{new FileInfo(chemin).Extension}");
+                    if (!File.Exists(imgNewPath) && File.Exists(imgNewPath))
                     {
                         File.Copy(chemin, imgNewPath);
-                    }
-                }
-                // Sous Dossiers
-                DataTable souDoss = dB.SousDossier(id_dossier);
-                foreach (DataRow sd in souDoss.Rows)
-                {
-                    string id_sd = sd["ID_SD"].ToString();
-                    string numSD = sd["NUMERO_SD"].ToString();
-                    string formaliteSD = sd["FORMALITE"].ToString();
-                    string sdName = $"sd{numSD} {formaliteSD}";
-                    string sdPath = Path.Combine(dossPath, sdName);
-                    if (Directory.Exists(sdPath))
-                    {
-                        Directory.CreateDirectory(sdPath);
-                    }
-
-
-                    // Pieces dans sous dossier
-                    DataTable pieces = dB.Piece(id_sd);
-                    foreach (DataRow piece in pieces.Rows)
-                    {
-                        string num_pg = piece["Numero_page"].ToString();
-                        string nomPiece = piece["Nature_Acte"].ToString();
-                        string chemin = piece["CHEMIN_PHYSIQUE"].ToString();
-                        string piecePath = Path.Combine(sdPath, nomPiece);
-                        if (!Directory.Exists(piecePath))
-                        {
-                            Directory.CreateDirectory(piecePath);
-                        }
-
-                        string imgNewPath = Path.Combine(piecePath, $"p{num_pg}{new FileInfo(chemin).Extension}");
-                        if (!File.Exists(imgNewPath))
-                        {
-                            File.Copy(chemin, imgNewPath);
-                        }
                     }
                 }
                 progress.Report(++i);
 
             }
 
+        }
+
+        private void ComboBaseDonnee_SelectedIndexChanged(object sender, EventArgs e)
+        {
+            var db = ComboBaseDonnee.Text;
+            conString = conString.Replace("master", db);
+
+
+            using (var con = new SqlConnection(conString))
+            {
+                con.Open();
+                string req = "select DISTINCT id_tranche from TB_UNITE";
+                SqlDataAdapter da = new SqlDataAdapter(req, con);
+                DataTable dt = new DataTable();
+                da.Fill(dt);
+                if (dt.Rows.Count == 0)
+                {
+                    MessageBox.Show("aucun tranche");
+
+                    return;
+                }
+                else
+                {
+                    comboxTranche.DataSource = dt;
+                    comboxTranche.DisplayMember = "id_tranche";
+                    comboxTranche.ValueMember = "id_tranche";
+                }
+            }
+        }
+
+        private void btnAnnuler_Click(object sender, EventArgs e)
+        {
+            this.Close();
         }
     }
 }
